+++ internal/templates/embed.go
@@ -1,0 +1,6 @@
+package templates
+
+import _ "embed"
+
+//go:embed files/execution_report.html
+var ExecutionReportTemplate string